pyproject.toml
[project]
name = "mi-autocot"
version = "0.1.0"
description = "MI-AutoCoT experiments with Hydra"
requires-python = ">=3.10"
dependencies = [
  "hydra-core>=1.3.2",
  "wandb>=0.16.0",
  "torch>=2.1.0",
  "transformers>=4.39.0",
  "datasets>=2.18.0",
  "sentence-transformers>=2.5.1",
  "scikit-learn>=1.3.0",
  "numpy>=1.24.0",
  "pandas>=2.1.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.13.0",
  "optuna>=3.5.0",
  "tqdm>=4.66.0",
  "accelerate>=0.26.0",
]
